pyproject.toml
[project]
name = "openlayers"
version = "0.1.0-rc.7"
description = "Python Bindings for OpenLayersJS"
readme = "README.md"
authors = [{ name = "Stefan Kuethe", email = "stefan.kuethe@eoda.de" }]
requires-python = ">=3.10, <4"
dependencies = [
    "anywidget>=0.9.0",
    "eval-type-backport>=0.2.2",
    "jinja2>=3.1.6",
    "pydantic>=2.5.3",
    "pyproj>=3.6.1",
]

# [project.scripts]
# openlayers = "openlayers:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "geopandas>=1.0.1",
    "isort>=6.0.1",
    "jupyter>=1.1.1",
    "marimo>=0.13.2",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.6.12",
    "mkdocstrings[python]>=0.29.1",
    "pyproj>=3.6.1",
    "pytest>=8.3.5",
    "requests>=2.32.3",
    "ruff>=0.11.5",
]
